% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignBPCellsObjs.R
\name{AlignBPCellsObjs}
\alias{AlignBPCellsObjs}
\title{Align BPCells Objects so they both have the same genes.}
\usage{
AlignBPCellsObjs(bpcells_obj1, bpcells_obj2)
}
\arguments{
\item{bpcells_obj1}{First BPCells Object}

\item{bpcells_obj2}{Second BPCells Object}
}
\value{
The first BPCells object put in the bpcells_obj1 parameter.
}
\description{
Align BPCells Objects so they both have the same genes.
}
\examples{
bpcells_obj1 = AlignBPCellsObjs(bpcells_obj1, bpcells_obj2)
bpcells_obj2 = AlignBPCellsObjs(bpcells_obj2, bpcells_obj1)
}
