% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_elastic_net.R
\name{binomial_elastic_net}
\alias{binomial_elastic_net}
\title{Binomial elastic net helper function}
\usage{
binomial_elastic_net(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
Binomial Elastic Net model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
Binomial elastic net helper function
}
\examples{
ex_model = binomial_elastic_net(reference_dataset, celltype_labels)
}
\keyword{internal}
