% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ridge.R
\name{binomial_ridge}
\alias{binomial_ridge}
\title{Binomial ridge helper function}
\usage{
binomial_ridge(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
Binomial ridge model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
Binomial ridge helper function
}
\examples{
data("iris")
dataset = iris[iris$Species \%in\% c("setosa","versicolor"),]
labels = as.factor(dataset$Species)
labels = droplevels(labels)
dataset = dataset[,1:4]
ex_model = binomial_ridge(dataset, labels)
}
\keyword{internal}
