% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEqualTree.R
\name{CreateEqualTree}
\alias{CreateEqualTree}
\title{This function is meant for users who don't want to use the tree-based/hierarchical functionality of the package.
It creates a tree that assumes that all the celltype labels are at the same level of the tree.}
\usage{
CreateEqualTree(cell_labels, rootnode_name = "Rootnode")
}
\arguments{
\item{cell_labels}{Vector of celltype labels we want to create a tree with. Don't include duplicate celltype labels.}

\item{rootnode_name}{Optional parameter giving the name of the rootnode of the tree.
Recommended to keep as "Unmapped".}
}
\value{
Tree in treedata format.
}
\description{
This function is meant for users who don't want to use the tree-based/hierarchical functionality of the package.
It creates a tree that assumes that all the celltype labels are at the same level of the tree.
}
\examples{
equal_tree = lionmap::CreateEqualTree(cell_labels = paste0(1:20,"n"),rootnode_name = "Root")
}
