% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetModels.R
\name{GetModels}
\alias{GetModels}
\title{Function that gives models trained to differentiate all pairwise matchups.}
\usage{
GetModels(
  marker_genes,
  ref_bpcells,
  ref_metadata,
  tree,
  metadata_cluster_column = "cluster_label",
  metadata_cell_id_column = "cell_label",
  n_cells_sampled = 500,
  models_to_include = NULL,
  npcs = 5
)
}
\arguments{
\item{marker_genes}{List with marker genes returned by the FindMarkerGenes function.}

\item{ref_bpcells}{BPCells obj with reference dataset GE values}

\item{ref_metadata}{Dataframe with metadata on each cell in reference dataset.}

\item{tree}{Tree structure in treedata format.}

\item{metadata_cluster_column}{Metadata celltype label column.}

\item{metadata_cell_id_column}{Metadata cell ID column}

\item{n_cells_sampled}{Number of cells used in pairwise model determination for each class.}

\item{models_to_include}{Optional vector which provides the names of models to include. If using
this parameter, include a subset of the following (make sure the names match or it won't work):
"linear_svm", "polynomial_svm", "naive_bayes", "ridge", "lasso", "elastic_net",
"linear_da", "knn", "rf", "quadratic_da"}

\item{npcs}{Optional parameter giving number of PCs to use in model creation.}
}
\value{
List of models that differentiates each pairwise matchup.
}
\description{
Function that gives models trained to differentiate all pairwise matchups.
}
\examples{
GetModels(marker_genes, ref_bpcells, ref_metadata, tree,
metadata_cluster_column = "cluster_label",
metadata_cell_id_column = "cell_label",
n_cells_sampled = 500, models_to_include = NULL, npcs = 5)
}
