% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{knn  helper function}
\usage{
knn(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
knn model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
knn  helper function
}
\examples{
ex_model = knn(reference_dataset, celltype_labels)
}
\keyword{internal}
