% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_da.R
\name{linear_da}
\alias{linear_da}
\title{linear_da  helper function}
\usage{
linear_da(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
linear_da model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
linear_da  helper function
}
\examples{
ex_model = linear_da(reference_dataset, celltype_labels)
}
