% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_svm.R
\name{linear_svm}
\alias{linear_svm}
\title{linear_svm  helper function}
\usage{
linear_svm(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
linear_svm model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
linear_svm  helper function
}
\examples{
data("iris")
dataset = iris[iris$Species \%in\% c("setosa","versicolor"),]
labels = as.factor(dataset$Species)
labels = droplevels(labels)
dataset = dataset[,1:4]
ex_model = linear_svm(dataset, labels)
}
\keyword{internal}
