% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lionmap-package.R
\docType{package}
\name{lionmap-package}
\alias{lionmap}
\alias{lionmap-package}
\title{lionmap: lionmap: tree-based single cell classification with algorithmic updates}
\description{
This package performs consensus classification on single cell datasets. Components are modular, and can be used standalone.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathan-columbiau/lionmap}
  \item \url{https://jonathan-columbiau.github.io/lionmap/}
  \item Report bugs at \url{https://github.com/jonathan-columbiau/lionmap/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Algoo \email{ja3527@cumc.columbia.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
