% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBPCellsFromTable.R
\name{LoadBPCellsFromTable}
\alias{LoadBPCellsFromTable}
\title{Helper function which uses the fread function from the data.table library to
read in a GE dataset in BPCells format.}
\usage{
LoadBPCellsFromTable(filename, dir_name = "new_dataset")
}
\arguments{
\item{filename}{Name of file}

\item{dir_name}{Name of BPCells obj}
}
\value{
BPCells-formatted GE matrix
}
\description{
Helper function which uses the fread function from the data.table library to
read in a GE dataset in BPCells format.
}
\examples{
LoadBPCellsFromTable(filename, dir_name = "bpcells_dataset")
}
\keyword{internal}
