% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBPCellsObj.R
\name{LoadBPCellsObj}
\alias{LoadBPCellsObj}
\title{Helper function that loads BPCells obj already created}
\usage{
LoadBPCellsObj(directory_name)
}
\arguments{
\item{directory_name}{Name of directory with BPCells Object.}
}
\value{
BPCells obj
}
\description{
Helper function that loads BPCells obj already created
}
\examples{
LoadBPCellsObj(bpcells_dir_name)
}
\keyword{internal}
