% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadBPCellsObjToSeurat.R
\name{LoadBPCellsObjToSeurat}
\alias{LoadBPCellsObjToSeurat}
\title{Helper function that takes in BPCells object and provides Seurat object
with this dataset (since Seurat also supports BPCells).}
\usage{
LoadBPCellsObjToSeurat(directory_name, metadata)
}
\arguments{
\item{directory_name}{Name of directory holding bpcells obj}

\item{metadata}{Dataframe of metadata for BPCells cells}
}
\value{
Seurat object with BPCells obj as GE matrix
}
\description{
Helper function that takes in BPCells object and provides Seurat object
with this dataset (since Seurat also supports BPCells).
}
\examples{
LoadBPCellsObjToSeurat(bpcells_dir, metadata_df)
}
\keyword{internal}
