% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_combinations.R
\name{pairwise_combinations}
\alias{pairwise_combinations}
\title{Helper function that takes in a vector of names, and provides a list with each pairwise combination
of names.}
\usage{
pairwise_combinations(cluster.names)
}
\arguments{
\item{cluster.names}{Vector of names/celltype labels}
}
\value{
List of pairwise combinations
}
\description{
Helper function that takes in a vector of names, and provides a list with each pairwise combination
of names.
}
\examples{
pairwise_combinations(names_vector)
}
\keyword{internal}
