% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_svm.R
\name{polynomial_svm}
\alias{polynomial_svm}
\title{polynomial_svm  helper function}
\usage{
polynomial_svm(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
polynomial_svm model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
polynomial_svm  helper function
}
\examples{
ex_model = polynomial_svm(reference_dataset, celltype_labels)
}
