% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_models.R
\name{predict_models}
\alias{predict_models}
\title{Helper function returning predicted classifications (formats and returns
consistent output forvarious packages)}
\usage{
predict_models(model, model_name, nonsparse_mat)
}
\arguments{
\item{model}{A model}

\item{model_name}{Name of model (must be one of the model names in package)}

\item{nonsparse_mat}{PCA-transformed matrix of query dataset}
}
\value{
Predicted Classifications
}
\description{
Helper function returning predicted classifications (formats and returns
consistent output forvarious packages)
}
\examples{
predict_models(lasso_model, "lasso", pca_transformed_matrix)
}
\keyword{internal}
