% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest}
\alias{random_forest}
\title{Random Forest  helper function (uses ranger backend)}
\usage{
random_forest(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
rf model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
Random Forest  helper function (uses ranger backend)
}
\examples{
data("iris")
dataset = iris[iris$Species \%in\% c("setosa","versicolor"),]
labels = as.factor(dataset$Species)
labels = droplevels(labels)
dataset = dataset[,1:4]
ex_model = random_forest(dataset, labels)
}
\keyword{internal}
