% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ex_data_bpcells}
\alias{test_ex_data_bpcells}
\title{PBMC 3k dataset subset}
\format{
\subsection{\code{test_ex_data_bpcells}}{

A bpcells object
}
}
\source{
\url{https://www.10xgenomics.com/datasets/3-k-pbm-cs-from-a-healthy-donor-1-standard-1-1-0}
}
\usage{
test_ex_data_bpcells
}
\description{
A subset of pbmc3k data stored in BPCells format used as test data in
vignette and examples.
}
\keyword{datasets}
