% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ex_metadata}
\alias{test_ex_metadata}
\title{PBMC 3k dataset testing metadata}
\format{
\subsection{\code{test_ex_metadata}}{

A bpcells object
}
}
\source{
\url{https://github.com/satijalab/seurat-data}
}
\usage{
test_ex_metadata
}
\description{
Metadata of pbmc3k dataset subset using Seurat annotations, taken from Seurat data
package.
}
\keyword{datasets}
