% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteSeuratToBPCellsDir.R
\name{WriteSeuratToBPCellsDir}
\alias{WriteSeuratToBPCellsDir}
\title{Helper function to write counts stored in Seurat object in the field
assays/RNA/counts to bpcells directory}
\usage{
WriteSeuratToBPCellsDir(seurat_obj, dir_name, cell_label_column = "cellid")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{dir_name}{BPCells directory name to write to}

\item{cell_label_column}{Cell label column name in Seurat object metadata}
}
\value{
BPCells GE matrix
}
\description{
Helper function to write counts stored in Seurat object in the field
assays/RNA/counts to bpcells directory
}
\examples{
WriteSeuratToBPCellsDir(seurat_obj, bpcells_dir, "cellid")
}
\keyword{internal}
