% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignBPCellsObjs.R
\name{AlignBPCellsObjs}
\alias{AlignBPCellsObjs}
\title{Align BPCells Objects so they both have the same genes (useful for getting reference and
query datasets in the proper format before finding marker genes).}
\usage{
AlignBPCellsObjs(bpcells_obj1, bpcells_obj2)
}
\arguments{
\item{bpcells_obj1}{First BPCells Object}

\item{bpcells_obj2}{Second BPCells Object}
}
\value{
The first BPCells object put in the bpcells_obj1 parameter.
}
\description{
Align BPCells Objects so they both have the same genes (useful for getting reference and
query datasets in the proper format before finding marker genes).
}
\examples{

train_ex_data_bpc = AlignBPCellsObjs(bpcells_obj1, bpcells_obj2)
bpcells_obj2 = AlignBPCellsObjs(bpcells_obj2, bpcells_obj1)
}
