% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_elastic_net.R
\name{binomial_elastic_net}
\alias{binomial_elastic_net}
\title{Binomial elastic net helper function}
\usage{
binomial_elastic_net(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
Binomial Elastic Net model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
Binomial elastic net helper function
}
\examples{
data("iris")
dataset = iris[iris$Species \%in\% c("setosa","versicolor"),]
labels = as.factor(dataset$Species)
labels = droplevels(labels)
dataset = dataset[,1:4]
ex_model = binomial_elastic_net(dataset, labels)
}
\keyword{internal}
