% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{Classify}
\alias{Classify}
\title{Classification function that takes in input dataset (in BPCells format), alongside
created models and the tree structure and provides a vector of classifications.}
\usage{
Classify(bpcells_query, models, tree_struc, prop_max_threshold = 0.66)
}
\arguments{
\item{bpcells_query}{Query dataset already aligned to reference dataset used to find marker genes and create models.}

\item{models}{Models created through GetMarkerGenes function.}

\item{tree_struc}{Tree used in model creation.}

\item{prop_max_threshold}{Proportion of evidence required for a test}
}
\value{
A vector providing classifications of cells in bpcells_query in the same order.
Performs specific unit testing on each run through.
Unit test 1: all remaining cells assigned to internal nodes
Unit test 2: expected number of elements returned
}
\description{
Classification function that takes in input dataset (in BPCells format), alongside
created models and the tree structure and provides a vector of classifications.
}
\examples{
Classify(query_obj, models, tree)
}
