% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHierarchy.R
\name{CreateHierarchy}
\alias{CreateHierarchy}
\title{Create Hierarchical Tree Structure from CSV with parent and child nodes
named in a csv file}
\usage{
CreateHierarchy(csv_file_path)
}
\arguments{
\item{csv_file_path}{A CSV file path with a column named 'parent' and a column
named 'node' giving the relationships between each node and its parent node.
This info will be used to create the tree.}
}
\value{
a treedata object
}
\description{
Create Hierarchical Tree Structure from CSV with parent and child nodes
named in a csv file
}
\examples{
CreateHierarchy("x.csv")
}
