% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes.R
\name{naive_bayes}
\alias{naive_bayes}
\title{naive_bayes  helper function}
\usage{
naive_bayes(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
naive_bayes model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
naive_bayes  helper function
}
\examples{
ex_model = naive_bayes(reference_dataset, celltype_labels)
}
