% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_da.R
\name{quadratic_da}
\alias{quadratic_da}
\title{Quadratic discriminant analysis  helper function}
\usage{
quadratic_da(reference_dataset, celltype_labels)
}
\arguments{
\item{reference_dataset}{PC-transformed dataset in matrix form}

\item{celltype_labels}{Celltype labels in vector}
}
\value{
quadratic_da model trained on evenly split dataset (upsamples if classes aren't evenly split)
}
\description{
Quadratic discriminant analysis  helper function
}
\examples{
ex_model = quadratic_da(reference_dataset, celltype_labels)
}
